+++ src/data_reliability/__init__.py
@@ -1,4 +1,13 @@
-from .core import DataTier, ReliableData, ReliabilityMetadata, ReliabilityPolicy
+from .core import DataTier, ReliableData, ReliabilityDecision, ReliabilityMetadata, ReliabilityPolicy
+from .database import (
+    RELIABILITY_COLUMNS,
+    metadata_from_columns,
+    metadata_from_document,
+    metadata_to_columns,
+    metadata_to_document,
+    scan_row,
+    trusted_records,
+)
 from .pandas_ext import filter_reliable_df
 from .scanner import (
     ReliabilityProfile,
@@ -15,6 +24,7 @@ from .scanner import (
 __all__ = [
     "DataTier",
     "ReliableData",
+    "ReliabilityDecision",
     "ReliabilityMetadata",
     "ReliabilityPolicy",
     "ReliabilityProfile",
@@ -22,9 +32,16 @@ __all__ = [
     "ReliabilityWeights",
     "TierCriterion",
     "ValidationEvidence",
+    "RELIABILITY_COLUMNS",
     "climate_record_profile",
     "compute_trace_hash",
     "default_profile",
     "filter_reliable_df",
+    "metadata_from_columns",
+    "metadata_from_document",
+    "metadata_to_columns",
+    "metadata_to_document",
+    "scan_row",
     "scientific_profile",
+    "trusted_records",
 ]